+++ src/bin/default.js
@@ -36,7 +36,7 @@ module.exports = {
   key: '',
   port: null,
   host: null,
-  deviceName: null,
+  // deviceName: null,
 
   // - - - - WEBDRIVER-IO  - - - -
   webdriverio: {