+++ HISTORY.md
@@ -21,9 +21,9 @@
       * that can be configured via the webdriverio API
         * [timeoutsAsyncScript](http://webdriver.io/api/protocol/timeoutsAsyncScript.html)
         * [timeoutsImplicitWait](http://webdriver.io/api/protocol/timeoutsImplicitWait.html)
-      
 * Fix: Output cucumber hook errors to the console
 * Fix: Saving screenshots
+* Fix: Use https:// for downloading selenium drivers (fixes #291)
 
 # 0.30.1